Create async credential validation service to avoid using injectors
